feat: Batch E — package exports, CLI, DM4 integration tests
- Update __init__.py exports to include PipelineExecutor, BatchProcessor,
  and __version__ = '0.2.0'
- Create examples/batch_workflow.py CLI entry point with argparse
- Create tests/test_integration_dm4.py for end-to-end DM4 testing
- Fix labeled_mask dtype from uint8 to uint16 to support >255 aerosols
  (discovered during real DM4 integration testing)

All 46 tests passing including DM4 integration tests.

+++ atem_analyzer/pipeline.py
@@ -98,7 +98,7 @@ class PipelineExecutor:
         """Generate reports for all detected aerosols."""
         # Labeled mask
         full_shape = original_signal.data.shape
-        labeled_mask = np.zeros(full_shape, dtype=np.uint8)
+        labeled_mask = np.zeros(full_shape, dtype=np.uint16)
         for i, obj in enumerate(aerosols):
             x, y, w, h = obj.bbox
             mask_global = np.zeros(full_shape, dtype=np.uint8)
